🔨 refactor(CardRepository.php): simplify create function and add input validation The create function has been simplified to remove the function arguments and instead retrieve the input values from the $_GET superglobal. Input validation has also been added to ensure that the input values are not null before being used in the query.
🐛 fix(CardRepository.php): update SQL query to exclude deleted cards
The SQL query has been updated to exclude deleted cards from the result set. This is achieved by adding a WHERE clause to the query that filters out any cards that have a non-null value in the deleted_at column.

🐛 fix(CardRepository.php): fix update method parameters
✨ feat(CardRepository.php): add support for updating card properties via GET parameters
The update method in CardRepository.php had incorrect parameters. The parameters have been removed and the card properties are now updated via GET parameters. The card properties are assigned to variables and then bound to the prepared statement. This allows for more flexibility in updating card properties.

🔨 refactor(edit.php): remove unused code and function
The `whatIsHappening()` function and related code have been removed as they are not used in the file. This improves the readability and maintainability of the code.

🔨 refactor(index.php): remove commented out code and unused variables
This commit removes the commented out code for the `whatIsHappening()` function and the call to it. It also removes the `$cards` and `$types` variables that are not used in the code.

🐛 fix(index.php): fix formatting and variable names
✨ feat(index.php): add edit functionality to update a card
The commit fixes the formatting and variable names in the index.php file. The create() function now calls the create() method of the CardRepository class without any parameters. The update() function now has two actions, "update" and "updated". When the action is "update", the function gets the specified card and requires the edit.php file. When the action is "updated", the function calls the update() method of the CardRepository class and redirects to the same page.

🔨 refactor(index.php): remove unnecessary global variables and add color array
The global variables $cards and $types are no longer needed as they are now being set within the function. The $types variable is now being set by calling the getTypes() method on the $cardRepository object. A new $colors array has been added to the function to be used later on.

🐛 fix(overview.php): remove unused variable and fix links
The variable $colorIndex is not used and can be removed. The links in the code were not working properly, so I removed the action attribute from the form and added a question mark to the href attributes of the links.